BREAK: move aws_lambda to aws but keep aws_lambda with deprecation warning
