+++ podpac/managers.py
@@ -4,4 +4,4 @@ Managers Public Module
 
 # REMINDER: update api docs (doc/source/user/api.rst) to reflect changes to this file
 
-from podpac.core.managers.aws_lambda import Lambda
+from podpac.core.managers.aws import Lambda
